insert_node

+++ hash_tables/3-hash_table_set.c
@@ -6,7 +6,7 @@
  */
 int hash_table_set(hash_table_t *ht, const char *key, const char *value)
 {
-	unsigned int add;
+	unsigned long int add;
 	hash_node_t *new = NULL;
 
 	new = malloc(sizeof(hash_node_t));
